requirements.txt
asyncpg==0.29.0
fastapi==0.111.0
pydantic==2.7.1
pydantic-settings==2.2.1
tortoise-orm==0.21.5
uvicorn==0.29.0
